Ticket [jri:#84 #84]: rename dos pacotes.
