+++ src/main/java/org/jrimum/utilix/ReadWriteStream.java
@@ -1,0 +1,71 @@
+/*
+ * Copyright 2008 JRimum Project
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at http://www.apache.org/licenses/LICENSE-2.0 Unless required by
+ * applicable law or agreed to in writing, software distributed under the
+ * License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS
+ * OF ANY KIND, either express or implied. See the License for the specific
+ * language governing permissions and limitations under the License.
+ * 
+ * Created at: 30/03/2008 - 18:17:10
+ * 
+ * ================================================================================
+ * 
+ * Direitos autorais 2008 JRimum Project
+ * 
+ * Licenciado sob a Licença Apache, Versão 2.0 ("LICENÇA"); você não pode usar
+ * esse arquivo exceto em conformidade com a esta LICENÇA. Você pode obter uma
+ * cópia desta LICENÇA em http://www.apache.org/licenses/LICENSE-2.0 A menos que
+ * haja exigência legal ou acordo por escrito, a distribuição de software sob
+ * esta LICENÇA se dará “COMO ESTÁ”, SEM GARANTIAS OU CONDIÇÕES DE QUALQUER
+ * TIPO, sejam expressas ou tácitas. Veja a LICENÇA para a redação específica a
+ * reger permissões e limitações sob esta LICENÇA.
+ * 
+ * Criado em: 30/03/2008 - 18:17:10
+ * 
+ */
+
+package org.jrimum.utilix;
+
+import java.io.Serializable;
+
+/**
+ * 
+ * <p>
+ * Interface genérica e representativa de coisas que necessitam de operações de
+ * fluxo de escrita e leitura.
+ * </p>
+ * 
+ * @author <a href=http://gilmatryx.googlepages.com/>Gilmar P.S.L.</a>
+ * 
+ * @since 0.2
+ * 
+ * @version 0.2
+ */
+
+public interface ReadWriteStream<G> extends Serializable {
+
+	/**
+	 * <p>
+	 * Escreve o tipo infomado.
+	 * </p>
+	 * 
+	 * @return Um valor no mesmo tipo do tipo parametrizado
+	 * @since 0.2
+	 */
+	public abstract G write();
+
+	/**
+	 * <p>
+	 * Lê o tipo informado.
+	 * </p>
+	 * 
+	 * @param g
+	 *            valor a ser lido
+	 * @since 0.2
+	 */
+	public abstract void read(G g);
+
+}
