+++ src/test/java/org/jrimum/utilix/TestPDFUtil.java
@@ -1,0 +1,74 @@
+/* 
+ * Copyright 2008 JRimum Project
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ * 
+ * Created at: 30/03/2008 - 23:49:00
+ *
+ * ================================================================================
+ *
+ * Direitos autorais 2008 JRimum Project
+ *
+ * Licenciado sob a Licença Apache, Versão 2.0 ("LICENÇA"); você não pode 
+ * usar esse arquivo exceto em conformidade com a esta LICENÇA. Você pode obter uma 
+ * cópia desta LICENÇA em http://www.apache.org/licenses/LICENSE-2.0 A menos que 
+ * haja exigência legal ou acordo por escrito, a distribuição de software sob esta 
+ * LICENÇA se dará “COMO ESTÁ”, SEM GARANTIAS OU CONDIÇÕES DE QUALQUER TIPO, sejam 
+ * expressas ou tácitas. Veja a LICENÇA para a redação específica a reger permissões 
+ * e limitações sob esta LICENÇA.
+ * 
+ * Criado em: 30/03/2008 - 23:49:00
+ * 
+ */
+
+package org.jrimum.utilix;
+
+import static org.junit.Assert.*;
+
+import org.junit.After;
+import org.junit.AfterClass;
+import org.junit.Before;
+import org.junit.BeforeClass;
+import org.junit.Test;
+
+public class TestPDFUtil {
+
+	@BeforeClass
+	public static void setUpBeforeClass() throws Exception {
+	}
+
+	@AfterClass
+	public static void tearDownAfterClass() throws Exception {
+	}
+
+	@Before
+	public void setUp() throws Exception {
+	}
+
+	@After
+	public void tearDown() throws Exception {
+	}
+
+	@Test
+	public void testChangeField2ImagePdfStamperFloatArrayImage() {
+		fail("Not yet implemented");
+	}
+
+	@Test
+	public void testChangeField2ImagePdfStamperRectanglePDFImage() {
+		fail("Not yet implemented");
+	}
+
+	@Test
+	public void testMergeFiles() {
+		fail("Not yet implemented");
+	}
+
+}
